fix: unify e2e test into one file
